feature(CalendarView): files organization
